Add Autoquant scaffold and CST
Set up autoquant scaffold. The autoquant functionality consists of
several steps. This commit introduces the general scaffold and first
preprocessing CST passes.

CST:
LibCST is used to obtain Concrete Syntax Trees. This commit introduces a
small number of passes that preprocess this AST before converting it
inot a CFG.

pysource:
The pysource submodule implements the higher level code generation,
i.e., it will take cake of package creation, class construction and
more. Currently, it is used to create a quantized class and add a
forward method.

quantops
The quantops submodule is a source of truth on quantized operations. It
implements an operator table that holds information on which arguments
should/could be quantized and which 'vanilla' operation it implements in
a quantized form. This table is used to perform function call and
operator replacement

Apply 8 suggestion(s) to 2 file(s)