Feature: skip unsupported quantizers in range setting
In cases where some quantizers support range setting, but other do not,
the range setting methods may fail because they expect all quantizers to
support range setting. This commit adds an option
`skip_unsupported_quantizers` to the current range setting methods that
will skip any non-supported quantizers. A log at the warning level is
written to the default logger to indicate that some quantizers of a
specific type are skipped.