Fix: move parameter initialization out of try block for param unpacking
Parameter initialization happened within the block that tries to unpack
variables in the `quantization_range` setter. As a result, any
`TypeError` or `ValueError` thrown durin initialization would be
reraised as an error related to the parameter unpacking. This doesn't
fix any error that causes the original error, but will ensure more
meaningful stack traces in the future.