Get rid of torch-onnx
torch-onnx is still requires for torch 2.4 but it will not be installed
as a dependency and must be installed manually.

Tests and `export` should use the same graph, otherwise an op might
be break down and become untraceable. For example the node
    %quantize_by_tile : [num_users=1] = call_function[target=torch.ops.fastforward.quantize_by_tile.default](args = (%x, %p_fc1_input_quantizer_scale, [32, 10], 8.0, torch.float32, %b_fc1_input_quantizer_offset), kwargs = {})
might be presend in the non decomposed graph as:
    %submod_49 : [num_users=1] = get_attr[target=submod_1]
    %quantize_by_tile : [num_users=1] = call_function[target=torch.ops.higher_order.wrap_with_set_grad_enabled](args = (False, %submod_49, %x, %p_fc1_input_quantizer_scale, %b_fc1_input_quantizer_offset), kwargs = {})
    %getitem : [num_users=1] = call_function[target=operator.getitem](args = (%quantize_by_tile, 0), kwargs = {})

Since the default decomposition table is used (not empty {}), `transpose` op is decomposed to permute op

torch 2.4+ asks to specify dim explicit
https://github.com/pytorch/pytorch/blob/v2.4.0/torch/nn/functional.py#L1819